Regex Search Visitor + event selector
* fixes #864
