+++ wegas-core/src/main/java/com/wegas/core/ejb/VariableDescriptorFacade.java
@@ -10,7 +10,7 @@ package com.wegas.core.ejb;
 import com.wegas.core.AlphanumericComparator;
 import com.wegas.core.Helper;
 import com.wegas.core.api.VariableDescriptorFacadeI;
-import com.wegas.core.ejb.GameModelFacade.FindAndReplaceVisitor;
+import com.wegas.core.tools.FindAndReplaceVisitor;
 import com.wegas.core.exception.client.WegasErrorMessage;
 import com.wegas.core.exception.internal.WegasNoResultException;
 import com.wegas.core.i18n.persistence.TranslatableContent;
@@ -41,7 +41,7 @@ import com.wegas.core.persistence.variable.scope.TeamScope;
 import com.wegas.core.persistence.variable.statemachine.DialogueDescriptor;
 import com.wegas.core.persistence.variable.statemachine.DialogueState;
 import com.wegas.core.persistence.variable.statemachine.DialogueTransition;
-import com.wegas.core.rest.FindAndReplacePayload;
+import com.wegas.core.tools.FindAndReplacePayload;
 import com.wegas.core.security.ejb.UserFacade;
 import com.wegas.mcq.ejb.QuestionDescriptorFacade;
 import com.wegas.mcq.persistence.ChoiceDescriptor;
